feat(utils): add a couple CVR-related functions
These will be useful for some upcoming work in VxAdmin.
