+++ libs/cvr-fixture-generator/src/utils.test.ts
@@ -1,4 +1,4 @@
-import { castVoteRecordHasWriteIns, generateCombinations } from './utils';
+import { generateCombinations } from './utils';
 
 test('generateCombinations', () => {
   expect(generateCombinations([1, 2, 3], 0)).toStrictEqual([]);
@@ -10,49 +10,4 @@ test('generateCombinations', () => {
   ]);
   expect(generateCombinations([1, 2, 3], 3)).toStrictEqual([[1, 2, 3]]);
   expect(generateCombinations([1, 2, 3], 4)).toStrictEqual([]);
-});
-
-test('castVoteRecordHasWriteIns with no votes', () => {
-  expect(
-    castVoteRecordHasWriteIns({
-      _ballotStyleId: '1',
-      _ballotType: 'standard',
-      _precinctId: '1',
-      _scannerId: '1',
-      _testBallot: false,
-      _batchId: '1',
-      _batchLabel: '1',
-    })
-  ).toEqual(false);
-});
-
-test('castVoteRecordHasWriteIns with non-write-in votes', () => {
-  expect(
-    castVoteRecordHasWriteIns({
-      _ballotStyleId: '1',
-      _ballotType: 'standard',
-      _precinctId: '1',
-      _scannerId: '1',
-      _testBallot: false,
-      _batchId: '1',
-      _batchLabel: '1',
-      mayor: ['mickey'],
-    })
-  ).toEqual(false);
-});
-
-test('castVoteRecordHasWriteIns with write-in votes', () => {
-  expect(
-    castVoteRecordHasWriteIns({
-      _ballotStyleId: '1',
-      _ballotType: 'standard',
-      _precinctId: '1',
-      _scannerId: '1',
-      _testBallot: false,
-      _batchId: '1',
-      _batchLabel: '1',
-      mayor: ['mickey'],
-      council: ['donald', 'write-in-0'],
-    })
-  ).toEqual(true);
 });
